updated feedback form based on suggestions
