+++ src/shared/ui/LinkButton.tsx
@@ -10,13 +10,16 @@ type Props = {
  */
 export default function LinkButton({ href, children, title }: React.PropsWithChildren<Props>) {
   return (
-    <a href={href} target="_blank" rel="noopener noreferrer" title={title ?? href}>
+    // added textDecoration to be none to avoid black underline when hovering over text on the button
+    <a href={href} target="_blank" rel="noopener noreferrer" title={title ?? href} style={{ textDecoration: 'none' }}>
       <button
         role="link"
         style={{
-          marginLeft: '0.5em',
-          marginBottom: '0.25em',
           padding: children !== '' ? '0.25em' : '0',
+          alignItems: 'center',
+          gap: '0.5em',
+          display: 'flex',
+          marginTop: '0.25em'
         }}
       >
         {children} <span aria-hidden={true}>↗</span>
